requirements.txt
asn1crypto==1.5.1
attrs==22.2.0
autopep8==2.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
clickclick==20.10.2
connexion==2.14.1
exceptiongroup==1.1.0
flake8==6.0.0
Flask==2.2.2
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.0.0
inflection==0.5.1
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.1
mccabe==0.7.0
packaging==23.0
pg8000==1.29.4
pluggy==1.0.0
pycodestyle==2.10.0
pyflakes==3.0.1
pyrsistent==0.19.3
pytest==7.2.1
python-dateutil==2.8.2
python-dotenv==0.21.1
PyYAML==6.0
requests==2.28.2
scramp==1.4.4
six==1.16.0
swagger-ui-bundle==0.0.9
tomli==2.0.1
urllib3==1.26.14
Werkzeug==2.2.2
zipp==3.11.0
